+++ src/domain/space/admin/SpaceCommunity/AdminOpportunityCommunityPage.tsx
@@ -2,12 +2,12 @@ import { FC } from 'react';
 import PageContent from '@/core/ui/content/PageContent';
 import PageContentBlock from '@/core/ui/content/PageContentBlock';
 import PageContentColumn from '@/core/ui/content/PageContentColumn';
-import CommunityOrganizations from '@/domain/community/community/CommunityAdmin/CommunityOrganizations';
+import CommunityOrganizations from '@/domain/space/admin/SpaceCommunity/CommunityOrganizations';
 import CommunityUsers from '@/domain/space/admin/SpaceCommunity/CommunityUsers';
 import useCommunityAdmin from '@/domain/space/admin/SpaceCommunity/useCommunityAdmin';
 import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLayout/SettingsSection';
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
-import CommunityVirtualContributors from '@/domain/community/community/CommunityAdmin/CommunityVirtualContributors';
+import CommunityVirtualContributors from '@/domain/space/admin/SpaceCommunity/CommunityVirtualContributors';
 import { SpaceLevel } from '@/core/apollo/generated/graphql-schema';
 import { useSubSpace } from '../../hooks/useSubSpace';
 import useUrlResolver from '@/main/routing/urlResolver/useUrlResolver';
