+++ src/domain/space/admin/SpaceCommunity/AdminSubspaceCommunityPage.tsx
@@ -4,15 +4,15 @@ import { Button } from '@mui/material';
 import PageContent from '@/core/ui/content/PageContent';
 import PageContentBlock from '@/core/ui/content/PageContentBlock';
 import PageContentColumn from '@/core/ui/content/PageContentColumn';
-import CommunityApplications from '@/domain/community/community/CommunityAdmin/CommunityApplications';
-import CommunityOrganizations from '@/domain/community/community/CommunityAdmin/CommunityOrganizations';
+import CommunityApplications from '@/domain/space/admin/SpaceCommunity/CommunityApplications';
+import CommunityOrganizations from '@/domain/space/admin/SpaceCommunity/CommunityOrganizations';
 import CommunityUsers from '@/domain/space/admin/SpaceCommunity/CommunityUsers';
 import useCommunityAdmin from '@/domain/space/admin/SpaceCommunity/useCommunityAdmin';
 import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLayout/SettingsSection';
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
 import { useSubSpace } from '../../hooks/useSubSpace';
 import InnovationLibraryIcon from '@/main/topLevelPages/InnovationLibraryPage/InnovationLibraryIcon';
-import CommunityVirtualContributors from '@/domain/community/community/CommunityAdmin/CommunityVirtualContributors';
+import CommunityVirtualContributors from '@/domain/space/admin/SpaceCommunity/CommunityVirtualContributors';
 import PageContentBlockSeamless from '@/core/ui/content/PageContentBlockSeamless';
 import InvitationOptionsBlock from '@/domain/community/invitations/InvitationOptionsBlock';
 import PageContentBlockCollapsible from '@/core/ui/content/PageContentBlockCollapsible';
