+++ src/domain/space/admin/SpaceCommunity/CommunityVirtualContributors.tsx
@@ -21,7 +21,7 @@ import { Actions } from '@/core/ui/actions/Actions';
 import { Identifiable } from '@/core/utils/Identifiable';
 import InviteVirtualContributorDialog from '@/domain/community/invitations/InviteVirtualContributorDialog';
 import { InviteContributorsData } from '@/domain/access/ApplicationsAndInvitations/useRoleSetApplicationsAndInvitations';
-import { ContributorViewProps } from '../EntityDashboardContributorsSection/Types';
+import { ContributorViewProps } from '../../../community/community/EntityDashboardContributorsSection/Types';
 import ButtonWithTooltip from '@/core/ui/button/ButtonWithTooltip';
 
 type RenderParams = GridRenderCellParams<string, ContributorViewProps>;
